Validate "original" task when processing results
As part of the performance improvements in d2b3b2c, we quit passing the
validated Task object from the TaskExecutor into the TaskResult to avoid
unnecessary copying (which can be very slow). Unfortunately this led to
using an object which may have template strings in fields. This patch
addresses that by making a shallow copy of the original task and post-
validating it for use in processing results. On average in my testing, this
added about 0.02s for each task on average, which should be an acceptible
performance hit compared to the original huge copy.

This patch also moves the handler tracking to be per-uuid, not per-object.
Doing it per-object had implications for the above due to the fact that the
copy of the original task is now being used, so the only sure way is to track
based on the uuid instead.

Fixes #18289